requirements.txt
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Statistical Models
statsmodels>=0.14.0
arch>=6.0.0  # For GARCH models

# Kalman Filtering
pykalman>=0.9.5
filterpy>=1.4.5

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Option Pricing
py_vollib>=1.0.1  # Black-Scholes and Greeks

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.23.0

# Configuration Management
pyyaml>=6.0

# Progress Bars (nice for beginners)
tqdm>=4.65.0

# Data Validation
pydantic>=2.0.0

# Logging
loguru>=0.7.0
